tracing: pull out span types into their own files
There's too much inter-mingling happening here which makes it a bit
tedious to pull out the various span types into their own packages. But
separate files makes it a bit easier to navigate, and can make do for
now.

(This commit is pure code-movement.)

Release note: None
